+++ backend/core/src/test/java/site/timecapsulearchive/core/common/fixture/MemberFixture.java
@@ -1,15 +1,14 @@
 package site.timecapsulearchive.core.common.fixture;
 
 import java.nio.charset.StandardCharsets;
+import site.timecapsulearchive.core.common.dependency.UnitTestDependency;
 import site.timecapsulearchive.core.domain.member.entity.Member;
 import site.timecapsulearchive.core.domain.member.entity.SocialType;
 import site.timecapsulearchive.core.global.security.encryption.HashEncryptionManager;
-import site.timecapsulearchive.core.global.security.encryption.HashProperties;
 
 public class MemberFixture {
 
-    private static final HashEncryptionManager hashEncryptionManager = new HashEncryptionManager(
-        new HashProperties("test"));
+    private static final HashEncryptionManager hashEncryptionManager = UnitTestDependency.hashEncryptionManager();
 
     public static Member member(int dataPrefix) {
         Member member = Member.builder()
